Add test handler interface.
